Add files via upload
I included the interpretations, but this is my first time working with an R mark down report so the formatting of the PDF is still wonky. Please check question 4 I used to chat for it and for questions three and four I included multiple outputs. I'm not sure which one I wanted to leave yet.